Patch removes an item from the items array in an order.
It then calls show and right now it attempts to make a synchronous API
call which executes before the update goes through.  This needs to be
changed to be made asynchronous.

The solution is to create an anonymous callback in a .then() which
calls the onShowOrder function inside events.

We created orderStore which separates concerns and contains the order that
is hardcoded into the showOrder function.  ALERT ALERT THIS NEEDS TO
CHANGE!

Tri-programmed with :
[Lauren](github.com/laurpaik) [Mario](github.com/mario7746)