Removes the embedded HTML form for stripe and replaces it with a javascript version of the same form which populates with the amount returned from the cart and some custom 0squad merch strings.
The checkout runs as a result of clicking the checkout button inside
the cart.

This makes a successful post to charges on the backend, which contains
a Stripe token that is proof of a successful purchase.

Right now this still adds the "pay with card" button to the cart,
which needs to be removed.