Include delete and index order features
Adds a div that shows the order history through an index API call.
Adds a delete button to each element in the order history.
Adds the destroy ajax function.
On successful destroy, refreshes the order index to reflect
the change.

Next step: Making the products load on page load.
Tri-programmed with Mario and Lauren