Allows users to abort checkout and add more items
Creates a global variable to store a current order ID

Uses that to check if an order has been posted or not, and then
if it has, it patches the new cart, and if it hasn't it posts
the new order.

This is able to happen because of a back end change that returns
the data so that we can display the updated order.

The UI function for clearCart now hides the modal and pops one extra
time to make sure there is nothing left in the cart.

Refactors the stripe events file so that the handler contains the
callback that causes the UI to change instead of the closed()
function.

Tri-programmed with Mario and Lauren