This successfully resets the orderComplete id.
Solved by turning it into an object that has a setter and a getter
as properties.  This allows the value in ID to persist between files.

Now after a successful checkout a new order begins.  The user can
still patch their order as before until they successfully checkout.

Tri programmed with Lauren and Mario